@@ -8,7 +8,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.3.1.RELEASE</version>
+        <version>3.3.5</version> 
         <relativePath/>
     </parent>
 
@@ -18,9 +18,9 @@
     <name>todo</name>
     <description>Demo project for Spring Boot</description>
 
-    <!-- On repasse en Java 8 (comme le projet d'origine) -->
     <properties>
-        <java.version>1.8</java.version>
+      
+        <java.version>23</java.version>
     </properties>
 
     <dependencies>
@@ -40,7 +40,6 @@
             <version>2.7.4</version>
         </dependency>
 
-        <!-- Lombok à jour et compatible (OK avec Java 8 aussi) -->
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
@@ -63,20 +62,18 @@
 
     <build>
         <plugins>
-            <!-- Plugin Spring Boot -->
             <plugin>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-maven-plugin</artifactId>
             </plugin>
 
-    
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.11.0</version>
                 <configuration>
-                    <source>${java.version}</source>
-                    <target>${java.version}</target>
+                    <source>${java.version}</source> 
+                    <target>${java.version}</target> 
                     <annotationProcessorPaths>
                         <path>
                             <groupId>org.projectlombok</groupId>
